+++ photutils/isophote/tests/test_fitter.py
@@ -8,7 +8,7 @@ import pytest
 from astropy.io import fits
 from numpy.testing import assert_allclose
 
-from photutils.datasets import get_path
+from photutils.datasets.load import _get_path
 from photutils.isophote.fitter import CentralEllipseFitter, EllipseFitter
 from photutils.isophote.geometry import EllipseGeometry
 from photutils.isophote.harmonics import fit_first_and_second_harmonics
@@ -150,8 +150,8 @@ def test_fitting_all():
 @pytest.mark.remote_data
 class TestM51:
     def setup_class(self):
-        path = get_path('isophote/M51.fits', location='photutils-datasets',
-                        cache=True)
+        path = _get_path('isophote/M51.fits', location='photutils-datasets',
+                         cache=True)
         hdu = fits.open(path)
         self.data = hdu[0].data
         hdu.close()